+++ src/main/java/parking/domain/models/Plazas.java
@@ -1,24 +1,21 @@
+
+
 package parking.domain.models;
 
 import jakarta.persistence.*;
 
 @Entity
 @Table(name = "Plazas")
-
 public class Plazas {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
+    private Integer id;
     private int numeroPlazas;
-
     private int sotano;
-
     private boolean disponible;
-
     private String propietario;
 
-    public Plazas(Long id, int numeroPlazas, int sotano, boolean disponible, String propietario) {
+    public Plazas(Integer id, int numeroPlazas, int sotano, boolean disponible, String propietario) {
         this.id = id;
         this.numeroPlazas = numeroPlazas;
         this.sotano = sotano;
@@ -29,11 +26,11 @@ public class Plazas {
     public Plazas() {
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
