+++ src/main/java/parking/domain/models/Usuarios.java
@@ -11,7 +11,7 @@ import jakarta.persistence.Table;
 public class Usuarios {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id; // Cambio de Long a Integer
 
     private String nombre;
     private String email;
@@ -30,11 +30,11 @@ public class Usuarios {
         this.rol = rol;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
