+++ src/main/java/parking/domain/services/SotanosService.java
@@ -1,10 +1,10 @@
 package parking.domain.services;
 
-import parking.domain.models.Sotanos;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.PersistenceContext;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import parking.domain.models.Sotanos;
 
 @Service
 public class SotanosService {
@@ -17,7 +17,7 @@ public class SotanosService {
     }
 
     @Transactional
-    public Sotanos obtenerSotanoPorId(Long id) {
+    public Sotanos obtenerSotanoPorId(Integer id) { // Cambio de Long a Integer
         return entityManager.find(Sotanos.class, id);
     }
 
@@ -27,7 +27,7 @@ public class SotanosService {
     }
 
     @Transactional
-    public void eliminarSotano(Long id) {
+    public void eliminarSotano(Integer id) { // Cambio de Long a Integer
         Sotanos sotano = obtenerSotanoPorId(id);
         if (sotano != null) {
             entityManager.remove(sotano);
