+++ src/main/java/parking/infraRepositories/EstadoPlazasRepository.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 import parking.domain.models.EstadoPlazas;
 
 @Repository
-public interface EstadoPlazasRepository extends JpaRepository<EstadoPlazas, Long> {
-    // Puedes agregar consultas personalizadas si es necesario
+public interface EstadoPlazasRepository extends JpaRepository<EstadoPlazas, Integer> {
+
 }
