+++ src/main/java/parking/infraRepositories/SotanosRepository.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 import parking.domain.models.Sotanos;
 
 @Repository
-public interface SotanosRepository extends JpaRepository<Sotanos, Long> {
-    // Puedes definir consultas personalizadas si es necesario
+public interface SotanosRepository extends JpaRepository<Sotanos, Integer> {
+
 }
